fix #5 and fix #14
