+++ DADMApp/app/src/main/java/com/example/dadmapp/DadmApp.kt
@@ -1,6 +1,8 @@
 package com.example.dadmapp
 
 import android.annotation.SuppressLint
+import android.os.Build
+import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
@@ -27,6 +29,7 @@ enum class RouteState(val title: String) {
     SignUp("SignUp")
 }
 
+@RequiresApi(Build.VERSION_CODES.TIRAMISU)
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -78,7 +81,9 @@ fun DadmApp(
             }
 
             composable("recordAudio") {
-                RecordAudioPage()
+                RecordAudioPage(
+                    onCreatedNote = { noteId -> navController.navigate("note/$noteId") }
+                )
             }
         }
     }
